// src/web3/types/listedToken.ts
//-----------------------------------------------------------------------------------------------//
import address from './address'


//-----------------------------------------------------------------------------------------------//
interface listedToken {
  "name": string
  "symbol": string
  "address": address
  "chainId": string
  "decimals": number
  "logo": string
}


//-----------------------------------------------------------------------------------------------//
export default listedToken


//-----------------------------------------------------------------------------------------------//